Move unit tests to separate module
